pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "task manager app"
authors = ["Yuliya Dzhumanazarova <yaschick451@gmail.com>"]
readme = "README.md"
packages = [{include = "hexlet_code"}]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"
django = "^4.2.3"
dj-database-url = "^2.0.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
django-bootstrap5 = "^23.3"
django-filter = "^23.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
